Implement provider redundancy (#62)
This change introduces redundancy in the alarms feed provider, allowing multiple instances to be deployed simultaneously such that if any one instance stops working, the remaining instances can be activated to take over the work of the failed instance.

This is accomplished by changing the feed actions to write directly to the trigger DB, rather than talking directly to the feed provider through REST endpoints. In this way, provider instances become aware of new and deleted triggers by monitoring the DB changes feed.